requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
python-magic==0.4.27
requests==2.31.0
urllib3==2.1.0
zammad-py==3.0.0
